requirements.txt
aiogram
asyncio
python-dotenv
xlrd
xlwt
sqlalchemy